refactor: modularize app concerns
