+++ src/settings.js
@@ -1,0 +1,3 @@
+export function initSettings() {
+  // Settings panel logic placeholder
+}
